+++ packages/elements/analytics-next-types/index.d.ts
@@ -1,6 +1,6 @@
 // For version "3.0.0"
 import * as React from 'react';
-import { PropsInjector } from '@atlaskit/type-helpers';
+import { PropsInjector, SumPropsInjector } from '@atlaskit/type-helpers';
 
 /*
   UIAnalyticsEvent.js
@@ -154,6 +154,11 @@ export type WithAnalyticsEventFunction = PropsInjector<WithAnalyticsEventProps>;
 export function withAnalyticsEvents<TOwnProps>(
   createEventMap?: EventMap<TOwnProps>,
 ): WithAnalyticsEventFunction;
+
+// Just in case your props are of Sum type, and not a Record.
+export function withAnalyticsForSumTypeProps<TOwnProps>(
+  createEventMap?: EventMap<TOwnProps>,
+): SumPropsInjector<WithAnalyticsEventProps>;
 
 /*
   createAndFireEvent.js