+++ packages/core/comment/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/comment
 
+## 6.0.35
+- Updated dependencies [0c116d6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/0c116d6):
+  - @atlaskit/editor-core@95.0.0
+
 ## 6.0.34
 - Updated dependencies [cbb8cb5](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/cbb8cb5):
   - @atlaskit/editor-core@94.0.0